ensure cv is installed even on live sites.

+++ etc/dockerfiles/powerbase/docker-entrypoint.sh
@@ -4,7 +4,7 @@ set -e
 # If the command passed to docker run/exec starts with a - or -- then
 # set the command.
 if [ "${1:0:1}" = '-' ]; then
-	set -- php-fpm7.0 "$@"
+  set -- php-fpm7.0 "$@"
 fi
 
 if [ "$1" = 'php-fpm7.0' ]; then
@@ -19,15 +19,13 @@ if [ "$LONG_LIVE" = "n" ]; then
   printf "html_errors = On\n" >> '/etc/php/7.0/fpm/conf.d/99-powerbase-dev.ini'
   printf "log_errors = On\n" >> '/etc/php/7.0/fpm/conf.d/99-powerbase-dev.ini'
 
-  # And we ensure the civix and cv is installed. Sometimes download.civicrm.org is not 
+  # And we ensure the civix is installed. Sometimes download.civicrm.org is not 
   # available - we don't want to hang forever, so timeout after 30 seconds.
-  set +e
-  wget --tries=2 --timeout=30 -O /usr/local/bin/civix https://download.civicrm.org/civix/civix.phar
-  wget --tries=2 --timeout=30 -O /usr/local/bin/cv https://download.civicrm.org/cv/cv.phar
-  set -e
-  chmod 755 /usr/local/bin/civix
-  chmod 755 /usr/local/bin/cv
-
+  if [ ! -f /usr/local/bin/civix ]; then
+    wget --tries=2 --timeout=30 -O /usr/local/bin/civix https://download.civicrm.org/civix/civix.phar
+    chmod 755 /usr/local/bin/civix
+  fi
+  
   # And activate  Jamie's private x509 cert
   if [ ! -f "/usr/local/share/ca-certificates/jamie.crt" ]; then
     mv /root/jamie.crt /usr/local/share/ca-certificates/
@@ -36,6 +34,12 @@ if [ "$LONG_LIVE" = "n" ]; then
 
   # And relay email to the smtp server, not bulk.mayfirst.org.
   sed -i "s/bulk\.ourpowerbase\.net/smtp/" /etc/esmtprc
+fi
+
+# We use cv to enable extensions.
+if [ ! -f /usr/local/bin/cv ]; then
+  wget --tries=2 --timeout=30 -O /usr/local/bin/cv https://download.civicrm.org/cv/cv.phar
+  chmod 755 /usr/local/bin/cv
 fi
 
 # Allow us to overwrite the DNS caching server at runtime.
